Tried to implement redux
